requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==4.4.2
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
moviepy==1.0.3
numpy==1.26.4
pillow==10.2.0
proglog==0.1.10
pytube3==9.6.4
requests==2.31.0
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1